requirements.txt
absl-py==2.3.1
accelerate==1.11.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
amqp==5.3.1
aniso8601==9.0.1
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
atomicwrites==1.4.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.12.3
bidict==0.23.1
billiard==4.2.1
binaryornot==0.4.4
bitsandbytes==0.48.1
bleach==6.2.0
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.0
celery==5.3.4
certifi==2024.2.2
cffi==1.17.1
channels==4.0.0
channels-redis==4.1.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.3
comtypes==1.4.8
constantly==23.10.4
contourpy==1.3.2
cookiecutter==2.6.0
crispy-bootstrap4==2024.10
crispy-tailwind==0.5.0
cron-descriptor==1.4.5
cryptography==43.0.3
cssselect2==0.8.0
curl_cffi==0.12.0
cycler==0.12.1
daphne==4.1.2
datasets==4.3.0
debugpy==1.8.17
decorator==5.2.1
deep-translator==1.11.4
Deprecated==1.2.18
dill==0.3.9
dj-database-url==2.3.0
Django==5.0.4
django-allauth==65.3.0
django-appconf==1.0.6
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cleanup==9.0.0
django-cors-headers==4.6.0
django-crispy-forms==2.1
django-environ==0.11.2
django-imagekit==5.0.0
django-tailwind==3.6.0
django-timezone-field==7.0
djangorestframework==3.15.1
dm-tree==0.1.9
dnspython==2.7.0
drf-spectacular==0.28.0
einops==0.8.1
email-validator==2.0.0
et_xmlfile==2.0.0
eventlet==0.33.3
executing==2.2.1
Faker==33.1.0
fastapi==0.120.2
filelock==3.20.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-RESTful==0.3.9
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
flatbuffers==25.9.23
fonttools==4.58.4
fpdf==1.7.2
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2025.9.0
gast==0.6.0
google-api-core==2.23.0
google-api-python-client==2.86.0
google-auth==2.36.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
google-trans-new==1.1.9
googleapis-common-protos==1.66.0
googletrans==3.1.0a0
greenlet==3.1.1
grpcio==1.76.0
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
h5py==3.15.1
hpack==3.0.0
hstspreload==2024.12.1
html5lib==1.1
httpcore==0.9.1
httplib2==0.22.0
httptools==0.6.4
httpx==0.13.3
huggingface-hub==0.36.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==24.7.2
inflection==0.5.1
iniconfig==2.3.0
ipykernel==7.0.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kaggle==1.7.4.5
kagglehub==0.3.13
keras==3.12.0
kiwisolver==1.4.8
kombu==5.4.2
langdetect==1.0.9
langid==1.1.6
libclang==18.1.1
limits==5.4.0
llvmlite==0.45.1
lxml==5.3.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistral_common==1.8.5
ml_dtypes==0.5.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
multitasking==0.0.11
mysql-connector==2.2.9
mysql-connector-python==9.1.0
namex==0.1.0
narwhals==2.10.1
nest-asyncio==1.6.0
networkx==3.5
numba==0.62.1
numpy==2.2.6
oauthlib==3.2.2
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.17.0
packaging==24.2
pandas==2.2.3
parso==0.8.5
pathos==0.3.3
peewee==3.18.2
peft==0.17.1
pilkit==3.0
pillow==11.0.0
platformdirs==4.3.8
plotly==6.3.1
pluggy==1.6.0
pox==0.3.5
ppft==1.7.6.9
prompt_toolkit==3.0.48
propcache==0.3.0
proto-plus==1.25.0
protobuf==5.29.0
psutil==7.1.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
py==1.11.0
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycountry==24.6.1
pycparser==2.22
pydantic==2.10.3
pydantic-extra-types==2.10.6
pydantic_core==2.27.1
pydyf==0.11.0
PyExecJS==1.5.1
Pygments==2.18.0
pyOpenSSL==24.2.1
pyotp==2.9.0
pyparsing==3.2.0
pyphen==0.17.2
pypiwin32==223
pytesseract==0.3.10
pytest==7.1.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.2.0
python-dotenv==1.0.0
python-engineio==4.8.0
python-magic==0.4.27
python-multipart==0.0.20
python-slugify==8.0.4
python-socketio==5.10.0
pyttsx3==2.98
pytz==2023.3.post1
pywin32==308
PyYAML==6.0.3
pyzmq==27.1.0
RapidFuzz==3.10.1
redis==5.0.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.5
requests-oauthlib==2.0.0
rfc3986==1.5.0
rich==13.9.4
rpds-py==0.27.1
rsa==4.9
safetensors==0.6.2
schedule==1.1.0
scikit-learn==1.7.2
scipy==1.14.1
seaborn==0.13.2
sentencepiece==0.2.1
service-identity==24.2.0
setuptools==80.9.0
shap==0.49.1
simple-websocket==1.1.0
six==1.16.0
slicer==0.0.8
slowapi==0.1.9
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
sqlparse==0.5.0
stack-data==0.6.3
starlette==0.47.2
stripe==7.6.0
sympy==1.14.0
syncedlyrics==1.0.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
tensorflow-probability==0.25.0
termcolor==3.2.0
text-unidecode==1.3
threadpoolctl==3.5.0
tiktoken==0.12.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.22.1
tomli==2.3.0
torch==2.9.0
tornado==6.5.2
tqdm==4.66.6
traitlets==5.14.3
transformers==4.57.1
translators==5.9.3
Twisted==24.10.0
txaio==23.1.1
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.38.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.13
weasyprint==66.0
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
whitenoise==6.6.0
wrapt==1.17.2
wsproto==1.2.0
WTForms==3.2.1
xgboost==3.1.1
xlrd==2.0.2
xxhash==3.6.0
yarl==1.18.3
yfinance==0.2.65
zope.interface==7.1.1
zopfli==0.2.3.post1
